iprt: RTPath - lenths are size_t not unsigned.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@14052 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/path.cpp
@@ -778,7 +778,7 @@ RTDECL(int) RTPathSharedLibs(char *pszPath, size_t cchPath)
  * @param   pszPath     Buffer where to store the path.
  * @param   cchPath     Buffer size in bytes.
  */
-RTDECL(int) RTPathAppDocs(char *pszPath, unsigned cchPath)
+RTDECL(int) RTPathAppDocs(char *pszPath, size_t cchPath)
 {
 #if !defined(RT_OS_WINDOWS) && defined(RTPATH_APP_DOCS)
     char *pszUtf8Path;
